// src/api/generated/models/gradeDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GradeDtoStatus } from './gradeDtoStatus';

export interface GradeDto {
  id?: string;
  companyId?: string;
  name?: string;
  levelOrder?: number;
  status?: GradeDtoStatus;
}